refactor: remove old chart component and enhance pie legend
